ia_engine/requirements.txt
fastapi==0.115.0
uvicorn==0.27.0
transformers==4.40.0
torch==2.3.0
sentence-transformers==2.2.2
faiss-cpu==1.8.0
pydub==0.25.1